+++ src/core/hotkey_utils.cpp
@@ -39,11 +39,4 @@ const char* VirtualKeyToString(int vkCode) {
     }
 }
 
-std::string FormatHotkeyConfig(int toggleKey, int recenterKey) {
-    std::ostringstream ss;
-    ss << VirtualKeyToString(toggleKey) << "=Toggle, ";
-    ss << VirtualKeyToString(recenterKey) << "=Recenter";
-    return ss.str();
-}
-
 } // namespace SkyrimHT